Content List

R6Class for storing content in a list. Note that this$container() and this$set_session(session) must be called in order for the content list to work properly.

Usage

content_list <- ContentList$new("content")

Methods

new(id, sortable = FALSE)

Initialize the content list.

id	Character. The id of the content list's container. If the session associated with the content list is a session proxy (session inside of a module), the id gets namespaced.

sortable	Currently defunct.

container()

Call this function in the UI definition. This function returns the div-element containing the content list.

set_session(session)

Call this function in the server function to connect the content list with a session.

session	A shiny session object.

add_element(content_element)

Add a content element to the content list. If this particular element has already been added and it is currently hidden, it gets shown.

content_element	A content element object.

add_element_actionButton(content_element, actionButton_id, actionButton_session)

Add a content element to the content list, that gets displayed after an actionButton has been clicked.

content_element	A content element object.

actionButton_id	The inputId associated with the actionButton.

actionButton_session	The shiny session object that the actionButton belongs to.

add_element_by_id(content_element_id)

Show the content element with id == content_element_id.

append_tab(content_element_id, tab, select = FALSE)

Append a tab to the content element with id == content_element_id.

content_element_id	Character. The id of the content element.

tab	The item to be added (must be created with tabPanel).

select	If TRUE tab gets selected upon being inserted.

get_content_element_by_id(content_element_id)

Get the content element with id == content_element_id.

get_content_element_ids()

Get the ids of the content elements as a character vector.

hide_content_element_by_id(content_element_id)

Hide the content element with id == content_element_id.

update_tab(content_element_id, selected)

Update the selected tab of the content element with id == content_element_id

.