FacDesign

R6Class representing a factorial design.

Usage

fac_design <- FacDesign$new(name, k)

Methods

new(name, k, rep = 1, center = 0, fac_names = LETTERS[seq_len(k)], response_name = "response", index_name = "index", response = NULL)

Initialize the factorial design.

name	Character. The factorial design's name

k	Integer. Number of variable factors.

center	Integer. Number of center points.

fac_names	Character vector containing the names of the variable factors.

response_name	Character. Name of the response variable.

index_name	Character. Name of the index variable.

response	Numeric vector containing the values of the response variable for the standardised factorial design. If response = NULL, you have to add a response later via this$add_response().

add_response(response)

Add or update the response of the factorial design.

response	Numeric vector containing the values of the response variable for the standardised factorial design.

get_name()

Get the factorial design's name.

get_names()

Get a list with the following entries:

fac	The variable factors' names.

index	The index' name.

name	The factorial design's name.

response	The factorial design response's name.

The arguments decide wheter a name is part of the list or not.

get_table(randomized = TRUE, index = TRUE, factors = TRUE, response = TRUE, center = FALSE)

Get a tibble representing the factorial design.

randomized	If TRUE, the rows are ordered in run order and not in standardised order.

index	If TRUE, the index column is included in the output.

factors	If TRUE, the factor columns are included in the output.

response	If TRUE, the response column is included in the output.

center	If TRUE, the rows containing observations for the center points are included in the output.

has_response()

Returns a logical indicating whether the factorial design has a response or not.

nrow()

Get the number of rows of the factorial design.

rename_fac_names(new_names, old_names = NULL)

Set the factors' names. If old_names = NULL the length of new_names must be equal to the number of factors, otherwise the length of new_names and old_names has to be equal.

new_names	Character vector.

old_names	Character vector or NULL.

rename_index_name(new_name)

Rename the index with the character new_name.

rename_name(new_name)

Rename the factorial design with the character new_name.

rename_response_name(new_name)

Rename the response with the character new_name.