Create a themed actionButton

Create a themed actionButton

actionButtonQW(inputId, label, icon = NULL, style = "material-flat",
  color = "default", size = "xs", block = FALSE, no_outline = TRUE,
  tooltip = NULL, dropdown = FALSE)

Arguments

inputId
The input slot that will be used to access the value.

label
The contents of the button or link-usually a text label, but you could also use any other HTML, like an image.

icon
An optional icon to appear on the button.

style
Style of the button, to choose between simple, bordered, minimal, stretch, jelly, gradient, fill, material-circle, material-flat, pill, float, unite.

color
Color of the button: default, primary, warning, danger, success, royal.

size
Size of the button: xs, sm, md, lg.

block
Logical, full width button.

no_outline
Logical, don't show outline when navigating with keyboard/ interact using mouse or touch.

tooltip
Text to appear as a tooltip on the button.

dropdown
Logical, button displayed in dropdown.