Collapsible tab box

Create a collapsible tabBox.

collapsible_tabBox(..., id = NULL, selected = NULL, title = NULL,
  width = 6, height = NULL, side = c("left", "right"))

Arguments

...
tabPanel elements to include in the tabset

id
If provided, you can use input$id in your server logic to determine which of the current tabs is active. The value will correspond to the value argument that is passed to tabPanel.

selected
The value (or, if none was supplied, the title) of the tab that should be selected by default. If NULL, the first tab will be selected.

title
Title for the tabBox.

width
The width of the box, using the Bootstrap grid system. This is used for row-based layouts. The overall width of a region is 12, so the default valueBox width of 4 occupies 1/3 of that width. For column-based layouts, use NULL for the width; the width is set by the column that contains the box.

height
The height of a box, in pixels or other CSS unit. By default the height scales automatically with the content.

side
Which side of the box the tabs should be on ("left" or "right"). When side="right", the order of tabs will be reversed.

closeable
Only functional in tabList_R6.

Details

The collapsible_tabBox has class "collapsible-tab-box", the enclosing div of the closing actionButton has class "div-btn-close".